Remove "Launch at start" option
This option didn't work anyways since the program didn't get the
required permissions when booted this way.

Instead users will have to add a shortcut to the program into their
startup folder. I might look in to how installers work to see if it
could be done with an installer instead.