Move the contents of the window to its own page
This will make it easier to create a page for typing in your coordinates
during the first time run.

Ideally I want to get the coords from an adress, but haven't found an
API for this yet, so for now the user must look up their own
coordinates, theres plenty of free online tools for this.

This prompted a bug with the config file for DayNightPapers. For some
reason the wp sections were no longer recognized, however advancing the
assembly version fixed this by forcing the compiler to re-evaluated the
config.