Add first time setup page
This will show a dialog to type in your coords.

Works okay atm, but if you type in the digits incorrectly (i.e. 8,5
instead of 8.5) or vice versa for your culture you'll end up with
invalid coordinates.

For now I've "fixed" this by trying to fetch sunrise time and catching
the exception in main thread, however this isn't the best idea since it
will block the main thread till it gets an answer and it will only query
the server the first time it's called. So it needs to be revised,
however it works "well enough" for now.