Show wallpapers in UI
Looks nice, and means that you can see your night paper during the day and vice versa.

The program would crash at first because it got an error when trying to delete the old image in cache because it was in use by the UI, in order to fix this I had to set cache options to OnLoad on the BitmapImage object, and set Image.Source to null before trying to load the new image into cache.