Set Topmost to false after window has been brought to front
Otherwiser it will be always on top.

I've found that the need to do this isn't nececary on single monitor setups since it will bring it to front automatically, however when I'm using my multi-monitor setup it won't do this if it's on other monitors however if you call topmost it will.

This seems like a weird way to do this but I have yet to find a better way.